pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pydantic",
    "pydantic-core",
    "typing-inspection",
    "typing_extensions",
    "sniffio",
    "anyio",
    "anthropic",
    "python-dotenv",
    "ollama",
    "arxiv",
    "mcp",
    "nest-asyncio",
    "openai>=1.101.0",
    "google-genai>=1.38.0",
    "google-generativeai>=0.8.5",
]
